+++ example/lib/main.dart
@@ -1,7 +1,7 @@
-import 'dart:convert';
 
+import 'package:example/pages/event_fetch_builder.page.dart';
+import 'package:example/pages/fetch_builder.page.dart';
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
 import 'package:fetcher/fetcher.dart';
 
 void main() {
@@ -39,17 +39,7 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  final _fetchController = ParameterizedFetchBuilderController<bool, String>();
-
-  bool withError = false;
-  bool dataClear = false;
-
-  Future<String> fetchTask(bool? withError) async {
-    final response = await http.get(Uri.parse('http://worldtimeapi.org/api/timezone/Europe/Paris'));
-    await Future.delayed(const Duration(seconds: 2));
-    if (withError == true) throw Exception('Error !');
-    return json.decode(response.body)['datetime'];
-  }
+  int currentPageIndex = 0;
 
   @override
   Widget build(BuildContext context) {
@@ -57,89 +47,28 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: const Text('Fetcher Example'),
       ),
-      body: Column(
-        children: [
-
-          // Header
-          const Padding(
-            padding: EdgeInsets.all(10),
-            child: Text('Fetcher with parameters'),
+      bottomNavigationBar: NavigationBar(
+        onDestinationSelected: (int index) {
+          setState(() {
+            currentPageIndex = index;
+          });
+        },
+        selectedIndex: currentPageIndex,
+        destinations: const <Widget>[
+          NavigationDestination(
+            icon: Icon(Icons.download),
+            label: 'FetchBuilder',
           ),
-
-          // Settings
-          CheckboxListTile(
-            title: const Text('Clear data first'),
-            value: dataClear,
-            onChanged: (value) {
-              setState(() {
-                dataClear = value!;
-              });
-            },
-          ),
-          CheckboxListTile(
-            title: const Text('With error'),
-            value: withError,
-            onChanged: (value) {
-              setState(() {
-                withError = value!;
-              });
-            },
-          ),
-
-          // Buttons
-          ElevatedButton(
-            onPressed: () => _fetchController.refresh(clearDataFirst: dataClear, param: withError),
-            child: const Text('Refresh'),
-          ),
-
-          // Fetcher
-          Expanded(
-            child: FetchBuilder<bool, String>.parameterized(
-              controller: _fetchController,
-              task: fetchTask,
-              builder: (context, data) {
-                return Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    Text(
-                      'Data is fetched :',
-                      style: Theme.of(context).textTheme.titleMedium,
-                      textAlign: TextAlign.center,
-                    ),
-                    Text(
-                      data,
-                      textAlign: TextAlign.center,
-                    ),
-                  ],
-                );
-              },
-            ),
-          ),
-          const Separator(),
-          const Padding(
-            padding: EdgeInsets.all(10),
-            child: Text('Dense Fetcher with Error'),
-          ),
-          FetchBuilder.basic<String>(
-            task: () async => throw Exception('Error !!'),
-            isDense: true,
-            builder: (context, data) => throw StateError('Should never reach this code'),
+          NavigationDestination(
+            icon: Icon(Icons.sync),
+            label: 'EventFetchBuilder',
           ),
         ],
       ),
-    );
-  }
-}
-
-class Separator extends StatelessWidget {
-  const Separator({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 1,
-      color: Colors.grey,
+      body: <Widget>[
+        const FetchBuilderPage(),
+        const EventFetchBuilderPage(),
+      ][currentPageIndex],
     );
   }
 }