+++ lib/fetcher.dart
@@ -3,6 +3,9 @@ library fetcher;
 export 'src/default_fetcher_config.dart';
 export 'src/fetcher_config.dart';
 export 'src/fetch_builder.dart';
+export 'src/event_fetch_builder.dart';
 export 'src/async_task_builder.dart';
 export 'src/widgets/async_form.dart';
 export 'src/utils.dart';
+
+export 'package:value_stream/value_stream.dart';