+++ lib/src/exceptions/fetch_exception.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/foundation.dart';
+
+class FetchException {
+  const FetchException(this.innerException, this.retry);
+
+  final Object innerException;
+  final VoidCallback retry;
+}